% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{berechnung_mischproben}
\alias{berechnung_mischproben}
\title{2-Wochenmischproben (bSaP) berechnen aus < 8 Tagesmischproben}
\usage{
berechnung_mischproben(mv_daten_df, bSaP = FALSE, startdatum_bSaP = NULL)
}
\arguments{
\item{mv_daten_df}{Tibble mit eingelesenen und vorprozessierten mv Daten}

\item{bSaP}{Logisch (Vorgabe \code{FALSE}). Sollen berechnete Mischproben verwendet werden?}

\item{startdatum_bSaP}{optionaler Vektor mit den Startdaten der 3-/4-Tages-Mischproben, getrennt nach den im Datensatz enthaltenen Jahren. Format: startdatum <- c("dd.mm.yyyy","dd.mm.yyyy")}
}
\value{
Tibble mit den zeitproportional berechneten Konzentrationen der 2-Wochenmischproben, basierend auf den Tagesmischproben innerhalb des jeweiligen Intervalls.
}
\description{
Wird innerhalb der Funktion "einlesen_mv_awel" aufgerufen und berechnet (falls bSaP = TRUE) auf Basis der verdichteten 3- oder 4-Tagesmischproben zeitproportional
die Konzentrationen der entsprechenden theoretischen 2-Wochenmischproben. Der jaehrliche Beginn der verdichteten Probenahme kann entweder manuell eingegeben werden,
oder wird automatisch aufgrund des Startdatums der ersten Mischprobe, die kuerzer als 8 Tage gesammelt wurde, gesetzt. Faellt das pro Jahr geschaetzte Startdatum
der Verdichtung auf einen Tag vor dem 15.03.20XX, wird ein Hinweis auf einen moeglichen Fehler in der Intervallberechnung ausgegeben.
}
