% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{einlesen_mv_awel}
\alias{einlesen_mv_awel}
\title{MV-Daten des AWEL einlesen}
\usage{
einlesen_mv_awel(
  mv_daten_pfad,
  vsa_lookup_pfad = NULL,
  bafu_lookup_pfad = NULL,
  bSaP = FALSE,
  startdatum_bSaP = NULL
)
}
\arguments{
\item{mv_daten_pfad}{Pfad zur kommaseparierten Datei mit den MV-Daten (fuer mehrere Dateien als Vektor von Dateipfaden)}

\item{vsa_lookup_pfad}{Pfad zur VSA-Tabelle \code{Tab_Substanzen.xlsx} (optional)}

\item{bafu_lookup_pfad}{Pfad zum BAFU-Namen-Lookup (optional)}

\item{bSaP}{Logisch (Vorgabe \code{FALSE}). Sollen berechnete Mischproben verwendet werden?}

\item{startdatum_bSaP}{Optionaler Vektor mit den Startdaten der 3-/4-Tages-Mischproben, getrennt nach den im Datensatz enthaltenen Jahren. Format: startdatum <- c("dd.mm.yyyy","dd.mm.yyyy")}
}
\value{
Dataframe mit MV-Daten
}
\description{
Liest MV-Daten (LIMS: Exportfilter_DB_Daten_mvwizr) des AWEL ein, formatiert sie um und uebergibt sie fuer die weitere Verarbeitung direkt and die mvwizr Funktion "einlesen_mv_gbl" (wrapper-Funktion). Die Funktion ersetzt also "einlesen_mv_gbl" aus mvwizr, damit der AWEL LIMS-Export direkt verarbeitet werden kann.
Sollen berechnete Mischproben verwendet werden (bSaP = TRUE), werden Konzentrationen aus 3-/4-Tagesmischproben zeitproportional zu 2-Wochenmischproben verrechnet. Das jaehrliche Startdatum der verdichteten Probenahme kann optional als Vektor eingegeben werden.
}
\examples{
mv_daten_pfad <- system.file("extdata", "LIMSexport.txt", package = "mvwizrAWEL")
v.startdatum <- c("27.03.2023", "25.03.2024")

mvdaten <- einlesen_mv_awel(mv_daten_pfad = mv_daten_pfad,
                            bSaP = FALSE,
                            startdatum_bSaP = v.startdatum)

}
